# agentledger/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/agents/{id}/roi": {
      "get": {
        "operationId": "AgentsController_roi",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      }
    },
...
    "/v1/agents/{id}": {
      "get": {
        "operationId": "AgentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      },
      "patch": {
        "operationId": "AgentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      },
      "delete": {
        "operationId": "AgentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      }
    },
...
    "/v1/analytics/spend": {
      "get": {
        "operationId": "AnalyticsController_spend",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/allocation": {
      "get": {
        "operationId": "AnalyticsController_allocation",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dimension",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "team",
                "app",
                "agent",
                "user"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/model-mix": {
      "get": {
        "operationId": "AnalyticsController_modelMix",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/platform-spend": {
      "get": {
        "operationId": "AnalyticsController_platformSpend",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/burndown": {
      "get": {
        "operationId": "AnalyticsController_burndown",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "virtualKeyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/risk": {
      "get": {
        "operationId": "AnalyticsController_risk",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/unit-economics": {
      "get": {
        "operationId": "AnalyticsController_unitEconomics",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcomeType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minConfidence",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/roi": {
      "get": {
        "operationId": "AnalyticsController_roi",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcomeType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minConfidence",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/agent-economics": {
      "get": {
        "operationId": "AnalyticsController_agentEconomics",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/v1/analytics/focus-export": {
      "get": {
        "operationId": "AnalyticsController_focusExport",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv",
                "json"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/pilot-report": {
      "get": {
        "operationId": "AnalyticsController_pilotReport",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "json",
                "md"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/agents/{agentId}": {
      "get": {
        "operationId": "AnalyticsController_agentDetail",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/source-reconciliation": {
      "get": {
        "operationId": "AnalyticsController_sourceReconciliation",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "required": false,
            "in": "query",
            "description": "Optional team_id filter (cost/value/risk reads honor it; ignored elsewhere).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "AgentLedger Control-Plane API",
    "description": "Tenant-scoped control plane: resources, auth, and analytics over the MVs.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateAgentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "runtimeType": {
            "type": "string"
          },
          "ownerUserId": {
            "type": "string"
          },
          "dataAccessScope": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "connectedTools": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "approvalStatus": {
            "type": "string"
          },
          "riskPosture": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}